Move DynamicData from Experimental to Core
