+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/tests/Azure.AI.Language.Conversations.Tests.csproj
@@ -15,7 +15,7 @@
 
   <ItemGroup>
     <ProjectReference Include="$(AzureCoreTestFramework)" />
-    <ProjectReference Include="..\..\..\core\Azure.Core.Experimental\src\Azure.Core.Experimental.csproj" />
+    <ProjectReference Include="..\..\..\core\Azure.Core\src\Azure.Core.csproj" />
     <ProjectReference Include="..\src\Azure.AI.Language.Conversations.csproj" />
   </ItemGroup>
