+++ sdk/core/Azure.Core/src/DynamicData/DynamicCaseMapping.cs
