+++ sdk/core/Azure.Core/src/DynamicData/MutableJsonElement.WriteTo.cs
