+++ sdk/core/Azure.Core/src/Serialization/AzureCoreExtensions.cs
@@ -6,6 +6,8 @@ using System.Collections.Generic;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
+using Azure.Core.Dynamic;
+using Azure.Core.Json;
 using Azure.Core.Serialization;
 
 namespace Azure
@@ -56,6 +58,37 @@ namespace Azure
         {
             JsonElement element = data.ToObjectFromJson<JsonElement>();
             return element.GetObject();
+        }
+
+        /// <summary>
+        /// Return the content of the BinaryData as a dynamic type.
+        /// </summary>
+        public static dynamic ToDynamicFromJson(this BinaryData utf8Json)
+        {
+            return utf8Json.ToDynamicFromJson(DynamicDataOptions.Default);
+        }
+
+        /// <summary>
+        /// Return the content of the BinaryData as a dynamic type.
+        /// </summary>
+        public static dynamic ToDynamicFromJson(this BinaryData utf8Json, DynamicCaseMapping caseMapping, DynamicDateTimeHandling dateTimeHandling = DynamicDateTimeHandling.Rfc3339)
+        {
+            DynamicDataOptions options = new()
+            {
+                CaseMapping = caseMapping,
+                DateTimeHandling = dateTimeHandling
+            };
+
+            return utf8Json.ToDynamicFromJson(options);
+        }
+
+        /// <summary>
+        /// Return the content of the BinaryData as a dynamic type.
+        /// </summary>
+        public static dynamic ToDynamicFromJson(this BinaryData utf8Json, DynamicDataOptions options)
+        {
+            MutableJsonDocument mdoc = MutableJsonDocument.Parse(utf8Json, DynamicData.GetSerializerOptions(options));
+            return new DynamicData(mdoc.RootElement, options);
         }
 
         private static object? GetObject(in this JsonElement element)
