+++ sdk/core/Azure.Core/tests/public/Azure.Core.Tests.Public.csproj
@@ -11,7 +11,7 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="$(AzureCoreTestFramework)" />
-    <ProjectReference Include="..\..\src\Azure.Core.Experimental.csproj" />
+    <ProjectReference Include="..\..\src\Azure.Core.csproj" />
   </ItemGroup>
   
   <!-- Shared source from Azure.Core -->
