+++ sdk/core/Azure.Core/tests/public/JsonDataTestHelpers.cs
